+++ src/bitops.intrinsic.t.ml
@@ -1,0 +1,24 @@
+
+let [@inline always] count_set_bits_32 x =
+  Ocaml_intrinsics_kernel.Int32.count_set_bits x
+  |> Int32.to_int
+
+let [@inline always] count_set_bits_64 x =
+  Ocaml_intrinsics_kernel.Int64.count_set_bits x
+  |> Int64.to_int
+
+let [@inline always] count_trailing_zeros_32 x =
+  Ocaml_intrinsics_kernel.Int32.count_trailing_zeros_nonzero_arg x
+  |> Int32.to_int
+
+let [@inline always] count_trailing_zeros_64 x =
+  Ocaml_intrinsics_kernel.Int64.count_trailing_zeros_nonzero_arg x
+  |> Int64.to_int
+
+let [@inline always] count_leading_zeros_32 x =
+  Ocaml_intrinsics_kernel.Int32.count_leading_zeros_nonzero_arg x
+  |> Int32.to_int
+
+let [@inline always] count_leading_zeros_64 x =
+  Ocaml_intrinsics_kernel.Int64.count_leading_zeros_nonzero_arg x
+  |> Int64.to_int