+++ src/fast_bitvector.ml
@@ -11,6 +11,8 @@ let failwithf s = Printf.ksprintf failwith s
 module type Element = sig
   type t
 
+  val to_string : t -> string
+
   val bit_size : int
   val byte_size : int
   val shift : int
@@ -40,10 +42,14 @@ module type Element = sig
   val lognot : t -> t
 
   val count_set_bits : t -> int
+  val count_trailing_zeros : t -> int
+  val count_leading_zeros : t -> int
 end
 
 module Element_32 = struct
   include Int32
+
+  let to_string = Printf.sprintf "0x%08lx"
 
   let bit_size = 32
   let byte_size = 4
@@ -58,7 +64,9 @@ module Element_32 = struct
   let set b i v = set b (i*byte_size) v
   let get b i = get b (i*byte_size)
 
-  let count_set_bits = Popcount.count_set_bits_32
+  let count_set_bits = Bitops.count_set_bits_32
+  let count_trailing_zeros = Bitops.count_trailing_zeros_32
+  let count_leading_zeros = Bitops.count_leading_zeros_32
 
   let of_int i =
     logand (of_int i) max_int
@@ -66,6 +74,8 @@ end
 
 module Element_64 = struct
   include Int64
+
+  let to_string = Printf.sprintf "0x%016Lx"
 
   let bit_size = 64
   let byte_size = 8
@@ -80,7 +90,9 @@ module Element_64 = struct
   let set b i v = set b (i*byte_size) v
   let get b i = get b (i*byte_size)
 
-  let count_set_bits = Popcount.count_set_bits_64
+  let count_set_bits = Bitops.count_set_bits_64
+  let count_trailing_zeros = Bitops.count_trailing_zeros_64
+  let count_leading_zeros = Bitops.count_leading_zeros_64
 
   let of_int i =
     logand (of_int i) max_int
@@ -91,6 +103,8 @@ module Element = (val
                     then (module Element_32 : Element)
                     else (module Element_64 : Element)
                  )
+
+let _ = Element.to_string
 
 let length t =
   Element.get t 0
@@ -651,19 +665,6 @@ let append a b =
   Bitblit.bitblit ~src:b ~dst:t ~src_pos:0 ~dst_pos:length_a ~len:length_b;
   t 
 
-let [@inline always] fold ~init ~f t =
-  let length = length t in
-  let acc = ref init in
-  for i = 0 to pred length do
-    (* CR smuenzel: process word at a time *)
-    acc := f !acc (Unsafe.get t i)
-  done;
-  !acc
-
-let map t ~f =
-  (init [@inlined hint]) (length t)
-    ~f:(fun i -> f (Unsafe.get t i))
-
 open Sexplib0
 
 module type Bit_ordering_spec = sig
@@ -672,6 +673,8 @@ module type Bit_ordering_spec = sig
 end
 
 module Bit_ordering_conversion(Spec : Bit_ordering_spec) = struct
+  module Spec = Spec
+
   type nonrec t = t
 
   let to_string t =
@@ -754,6 +757,141 @@ module Bit_zero_last = struct
   let t_of_sexp = t_of_sexp
 end
 
+module type Fold_ordering_spec = sig
+  val get_index : length:int -> i:int -> int
+  
+  type ('a, 'b) fold_f
+  type ('a, 'b) foldi_f
+
+  val apply_fold_f : ('a, 'b) fold_f -> 'b -> 'a -> 'a
+  val apply_foldi_f : ('a, 'b) foldi_f -> int -> 'b -> 'a -> 'a
+
+  val zeros_until_next_bit : Element.t -> int
+  val skip_next_bit : Element.t -> int -> Element.t
+
+end
+
+module Fold_ordering = struct
+  (* For fold_left *)
+  module Bit_zero_first = struct
+    include Bit_zero_first'.Spec
+
+    type ('a, 'b) fold_f = 'a -> 'b -> 'a
+    type ('a, 'b) foldi_f = 'a -> int -> 'b -> 'a
+
+    let [@inline always] apply_fold_f f b acc = f acc b
+    let [@inline always] apply_foldi_f f i b acc = f acc i b
+
+    let zeros_until_next_bit e = Element.count_trailing_zeros e
+    let skip_next_bit = Element.shift_right_logical
+  end
+
+  (* For fold_right *)
+  module Bit_zero_last = struct
+    include Bit_zero_last'.Spec
+
+    type ('a, 'b) fold_f = 'b -> 'a -> 'a
+    type ('a, 'b) foldi_f = int -> 'b -> 'a -> 'a
+
+    let [@inline always] apply_fold_f f b acc = f b acc
+    let [@inline always] apply_foldi_f f i b acc = f i b acc
+
+    let zeros_until_next_bit e = Element.count_leading_zeros e
+    let skip_next_bit = Element.shift_left
+  end
+end
+
+module [@inline always] Basic_fold(Spec : Fold_ordering_spec) = struct
+  let [@inline always] foldi t ~init ~f =
+    let length = length t in
+    let acc = ref init in
+    for i = 0 to pred length do
+      acc := f !acc i (Unsafe.get t (Spec.get_index ~length ~i))
+    done;
+    !acc
+
+  let [@inline always] fold t ~init ~f =
+    foldi t ~init ~f:(fun acc _i b -> f acc b)
+
+  let [@inline always] foldi t ~init ~f =
+    foldi t ~init ~f:(fun acc i b -> Spec.apply_foldi_f f i b acc)
+
+  let [@inline always] fold t ~init ~f =
+    fold t ~init ~f:(fun acc b -> Spec.apply_fold_f f b acc)
+
+  let [@inline always] iteri t ~f =
+    let length = length t in
+    for i = 0 to pred length do
+      f i (Unsafe.get t (Spec.get_index ~length ~i))
+    done
+
+  let [@inline always] iter t ~f =
+    iteri t ~f:(fun _i b -> f b)
+
+  (* Only consider set bits *)
+  let [@inline always] fold_set t ~init ~f =
+    let length = length t in
+    let total_data_words = total_data_words ~length in
+    let acc = ref init in
+    for i = 1 to total_data_words do
+      let i' = Spec.get_index ~length:total_data_words ~i:(i - 1) + 1 in
+      let word = ref (Element.get t i') in
+      let subindex = ref 0 in
+      while Bool.not (Element.equal !word Element.zero) do
+        let tail = Spec.zeros_until_next_bit !word in
+        subindex := !subindex + tail;
+        word := Spec.skip_next_bit !word tail;
+        word := Spec.skip_next_bit !word 1;
+        let subindex' = Spec.get_index ~length:Element.bit_size ~i:!subindex in
+        let user_index = (i' - 1) * Element.bit_size + subindex' in
+        acc := Spec.apply_fold_f f user_index !acc;
+        incr subindex;
+      done
+    done;
+    !acc
+
+end
+
+module Fold_forward = struct 
+  include Basic_fold(Fold_ordering.Bit_zero_first)
+
+  let [@inline always] iter_set t ~f =
+    fold_set t ~init:() ~f:(fun () i -> f i)
+end
+
+module Fold_backward = struct
+  include Basic_fold(Fold_ordering.Bit_zero_last)
+
+  let [@inline always] iter_set t ~f =
+    fold_set t ~init:() ~f:(fun i () -> f i)
+end
+
+let fold = Fold_forward.fold
+let fold_left = Fold_forward.fold
+let foldi = Fold_forward.foldi
+let fold_lefti = Fold_forward.foldi
+let iter = Fold_forward.iter
+let iteri = Fold_forward.iteri
+let fold_set = Fold_forward.fold_set
+let fold_left_set = Fold_forward.fold_set
+let iter_set = Fold_forward.iter_set
+
+let fold_right = Fold_backward.fold
+let fold_righti = Fold_backward.foldi
+let rev_iter = Fold_backward.iter
+let rev_iteri = Fold_backward.iteri
+let rev_iter_set = Fold_backward.iter_set
+let fold_right_set = Fold_backward.fold_set
+
+let map t ~f =
+  (init [@inlined hint]) (length t)
+    ~f:(fun i -> f (Unsafe.get t i))
+
+let mapi t ~f =
+  (init [@inlined hint]) (length t)
+    ~f:(fun i -> f i (Unsafe.get t i))
+
+
 module Builder = struct
   type t =
     { mutable current : int
@@ -771,9 +909,9 @@ module Builder = struct
       match rev_list with
       | [] -> ()
       | x :: xs ->
-          let current_element_bit0 = (list_len - i - 1) * Sys.int_size in
-          Unsafe.With_int.or_ ~dst:bv bv ~bit0_at:current_element_bit0 x;
-          aux bv xs ~i:(i + 1)
+        let current_element_bit0 = (list_len - i - 1) * Sys.int_size in
+        Unsafe.With_int.or_ ~dst:bv bv ~bit0_at:current_element_bit0 x;
+        aux bv xs ~i:(i + 1)
     in
     aux bv t.rev_list ~i:0;
     if t.current_index > 0
@@ -818,4 +956,8 @@ module Builder = struct
 
   let of_iter iter = add_iter (create ()) iter
   let of_seq seq = add_seq (create ()) seq
+end
+
+module Private = struct
+  module Bitops = Bitops
 end